refactor(web): move server messages directory
